+++ src/main/java/net/forcemaster_rpg/ForcemasterClassMod.java
@@ -30,7 +30,7 @@ public class ForcemasterClassMod implements ModInitializer {
 
 
 	public static ConfigManager<ItemConfig> itemConfig = new ConfigManager<ItemConfig>
-			("items_v2", Default.itemConfig)
+			("items_v3", Default.itemConfig)
 			.builder()
 			.setDirectory(MOD_ID)
 			.sanitize(true)
@@ -44,7 +44,7 @@ public class ForcemasterClassMod implements ModInitializer {
 			.constrain(LootConfig::constrainValues)
 			.build();
 	public static ConfigManager<EffectsConfig> effectsConfig = new ConfigManager<EffectsConfig>
-			("effects", new EffectsConfig())
+			("effects_v1", new EffectsConfig())
 			.builder()
 			.setDirectory(MOD_ID)
 			.sanitize(true)