+++ src/main/java/net/forcemaster_rpg/item/weapons/WeaponsRegister.java
@@ -2,6 +2,8 @@ package net.forcemaster_rpg.item.weapons;
 
 import net.forcemaster_rpg.ForcemasterClassMod;
 import net.forcemaster_rpg.item.ForcemasterGroup;
+import net.minecraft.entity.attribute.EntityAttributeModifier;
+import net.minecraft.entity.attribute.EntityAttributes;
 import net.minecraft.item.Item;
 import net.minecraft.item.Items;
 import net.minecraft.item.ToolMaterials;
@@ -20,6 +22,13 @@ import java.util.function.Supplier;
 
 public class WeaponsRegister {
     public static final ArrayList<Weapon.Entry> entries = new ArrayList<>();
+
+    private static ItemConfig.Attribute armorAddition(float value) {
+        return new ItemConfig.Attribute(
+                "generic.armor",
+                value,
+                EntityAttributeModifier.Operation.ADDITION);
+    }
 
     private static Weapon.Entry entry(String name, Weapon.CustomMaterial material, Item item, ItemConfig.Weapon defaults) {
         return entry(null, name, material, item, defaults);
@@ -46,7 +55,7 @@ public class WeaponsRegister {
         };
     }
 
-    public static float knuckle_attackSpeed = -2.0f;
+    public static float knuckle_attackSpeed = -2.2f;
 
     //KNUCKLES
     private static Weapon.Entry knuckles(String name, Weapon.CustomMaterial material, float damage) {
@@ -91,7 +100,7 @@ public class WeaponsRegister {
             ;
     public static final Weapon.Entry netherite_knuckle = knuckles("netherite_knuckle",
             Weapon.CustomMaterial.matching(ToolMaterials.NETHERITE, () -> Ingredient.ofItems(Items.NETHERITE_INGOT)), 5.5F)
-            .attribute(ItemConfig.Attribute.bonus(SpellSchools.ARCANE.id, 4.5F))
+            .attribute(ItemConfig.Attribute.bonus(SpellSchools.ARCANE.id, 4.0F))
             .attribute(ItemConfig.Attribute.multiply(SpellPowerMechanics.HASTE.id, 0.05F))
             ;
     public static final Weapon.Entry legendary_golden_knuckle = knuckles("legendary_golden_knuckle",
@@ -102,8 +111,8 @@ public class WeaponsRegister {
             ;
     public static final Weapon.Entry guardian_knuckle = knuckles("guardian_knuckle",
             Weapon.CustomMaterial.matching(ToolMaterials.DIAMOND, () -> Ingredient.ofItems(Items.IRON_INGOT)),5.0F)
-            .attribute(ItemConfig.Attribute.bonus(SpellSchools.ARCANE.id, 3.5F))
-            .attribute(ItemConfig.Attribute.multiply(SpellPowerMechanics.CRITICAL_DAMAGE.id, 0.1F))
+            .attribute(ItemConfig.Attribute.bonus(SpellSchools.ARCANE.id, 3.0F))
+            .attribute(armorAddition(4.0F))
             .attribute(ItemConfig.Attribute.multiply(SpellPowerMechanics.HASTE.id, 0.025F))
             ;
     public static final Weapon.Entry bloody_knuckle = special_knuckle_1("bloody_knuckle",